requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
bottle==0.12.25
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
multidict==6.0.5
packaging==23.2
pycparser==2.21
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
yarl==1.9.4